AgentAsCode_BE/requirements.txt

chromadb
cryptography
ipython
Jinja2
langchain
langchain_chroma
langchain_community
langchain_core
langchain_experimental
langchain_openai
langgraph
numpy
pandas
Pillow
protobuf
psycopg_pool
python-dotenv
tornado
uvicorn
psycopg-pool
langgraph-checkpoint-postgres
sentence-transformers
tabulate
langchain_postgres
pyodbc
matplotlib
pycryptodome
psycopg2-binary
azure-storage-blob
python-multipart
google-adk==1.14.1
fastapi 
sqlalchemy
dotenv 
psycopg2
azure-storage-blob 
python-multipart
httpx
uvicorn 
fastmcp


pydantic-settings
scikit-learn